# back/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Fonosoft",
        "contact": {
            "name": "Mari\u00f3n Juan Pablo",
            "url": "https://www.example.com"
        },
        "version": "1.0"
    },
    "host": "fonosoft.local/",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Crear nuevo usuario",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Json format",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Json format",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Json format",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "formData",
                        "description": "Json format",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "\u00c9xito: un usuario reci\u00e9n creado",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "200": {
                        "description": "\u00c9xito: operaci\u00f3n exitosa"
                    },
                    "401": {
                        "description": "Rechazado: no autenticado"
                    },
                    "422": {
                        "description": "Falta campo obligatorio"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "registrar a un usuario",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Json format",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Json format",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "remember_me",
                        "in": "formData",
                        "description": "Json format",
                        "required": true,
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u00c9xito: operaci\u00f3n exitosa"
                    },
                    "401": {
                        "description": "Rechazado: no autenticado"
                    },
                    "422": {
                        "description": "Falta campo obligatorio"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Cerrar sesi\u00f3n de un usuario",
                "responses": {
                    "200": {
                        "description": "\u00c9xito: operaci\u00f3n exitosa"
                    },
                    "401": {
                        "description": "Rechazado: no autenticado"
                    },
                    "422": {
                        "description": "Falta campo obligatorio"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "405": {
                        "description": "Entrada inv\u00e1lida"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/user": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "devolver el usuario",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u00c9xito: operaci\u00f3n exitosa"
                    },
                    "401": {
                        "description": "Rechazado: no autenticado"
                    },
                    "422": {
                        "description": "Falta campo obligatorio"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "405": {
                        "description": "Entrada inv\u00e1lida"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "definitions": {
        "User": {
            "required": [
                "name",
                "email",
                "password",
                "password_confirmation"
            ],
            "properties": {
                "name": {
                    "description": "Nombre de usuario",
                    "type": "string",
                    "example": "John Conor"
                },
                "email": {
                    "description": "Direcci\u00f3n de correo electr\u00f3nico",
                    "type": "string",
                    "example": "john.conor@terminator.com"
                },
                "password": {
                    "description": "Una contrase\u00f1a muy segura",
                    "type": "string",
                    "example": "123456"
                },
                "password_confirmation": {
                    "description": "Confirmar contrase\u00f1a",
                    "type": "string",
                    "example": "123456"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "description": "Auth Bearer Token Format as 'Bearer <access_token>'",
            "name": "Authorization",
            "in": "header"
        }
    }
}